+++ packages/ripple-tide-landing-page/src/index.ts
@@ -26,6 +26,7 @@ const tideLandingPageModule: RplTideMapping = {
   schema: '@dpc-sdp/ripple-tide-landing-page/types',
   mapping: {
     ...tidePageBaseMapping({
+      withTopicTags: true,
       withSidebarSiteSectionNav: true,
       withSidebarContacts: true,
       withSidebarRelatedLinks: true,
@@ -34,6 +35,8 @@ const tideLandingPageModule: RplTideMapping = {
     summary: 'field_landing_page_summary',
     showHeroAcknowledgement: 'field_show_ack_of_country',
     showInPageNav: 'field_show_table_of_content',
+    showHeroImageCaption: 'field_show_hero_image_caption',
+    showTopicTags: 'field_show_topic_term_and_tags',
     inPageNavHeadingLevel: (src) => {
       if (src.field_node_display_headings === 'showH2AndH3') {
         return 'h3'
@@ -68,6 +71,7 @@ const tideLandingPageModule: RplTideMapping = {
   },
   includes: [
     ...tidePageBaseIncludes({
+      withTopicTags: true,
       withSidebarSiteSectionNav: true,
       withSidebarContacts: true,
       withSidebarRelatedLinks: true,