+++ packages/ripple-tide-landing-page/src/index.vue
@@ -1,5 +1,16 @@
 <template>
-  <RplLayout :background="page.background">
+  <TideBaseLayout
+    :site="site"
+    :background="page.background"
+    :pageTitle="page.title"
+    :pageDescription="page.description"
+    :pageLanguage="page.lang"
+    :footerImageCaption="
+      page.showHeroImageCaption ? page.heroHeader.backgroundImageCaption : ''
+    "
+    :topicTags="page.showTopicTags ? page.topicTags : []"
+    :updatedDate="page.changed || page.created"
+  >
     <template #aboveHeader>
       <slot name="aboveHeader"></slot>
     </template>
@@ -14,7 +25,10 @@
         :header="page.heroHeader"
         :hideBottomCornerGraphic="!!page.primaryCampaign"
       />
-      <TideLandingPageHeroAcknowledgement v-if="page.showHeroAcknowledgement" />
+      <TideLandingPageHeroAcknowledgement
+        v-if="page.showHeroAcknowledgement"
+        :message="site?.acknowledgementHeader"
+      />
       <TideDynamicComponents
         v-if="page.headerComponents?.length > 0"
         :components="page.headerComponents"
@@ -56,13 +70,14 @@
     <template #footer>
       <slot name="footer"> </slot>
     </template>
-  </RplLayout>
+  </TideBaseLayout>
 </template>
 
 <script setup lang="ts">
 import type { TideLandingPagePage } from './types'
 
 defineProps<{
+  site: any
   page: TideLandingPagePage
 }>()
 </script>