+++ packages/ripple-tide-publication/src/index.vue
@@ -3,7 +3,13 @@ export default { name: 'TidePublicationPage' }
 </script>
 
 <template>
-  <RplLayout :background="page.background">
+  <TideBaseLayout
+    :site="site"
+    :pageTitle="page.title"
+    :pageDescription="page.description"
+    :pageLanguage="page.lang"
+    :updatedDate="page.changed || page.created"
+  >
     <template #aboveHeader>
       <slot name="aboveHeader"></slot>
     </template>
@@ -39,13 +45,14 @@ export default { name: 'TidePublicationPage' }
     <template #footer>
       <slot name="footer"></slot>
     </template>
-  </RplLayout>
+  </TideBaseLayout>
 </template>
 
 <script setup lang="ts">
 import type { TidePublicationPage } from './types'
 
 interface Props {
+  site: any
   page: TidePublicationPage
 }